added reader and example
