+++ src/pypgoutput/__init__.py
@@ -1,14 +1,24 @@
-from pypgoutput.decoders import PgoutputMessage
-from pypgoutput.decoders import Begin
-from pypgoutput.decoders import Commit
-from pypgoutput.decoders import Origin
-from pypgoutput.decoders import Relation
-from pypgoutput.decoders import TupleData
-from pypgoutput.decoders import Insert
-from pypgoutput.decoders import Update
-from pypgoutput.decoders import Delete
-from pypgoutput.decoders import Truncate
-from pypgoutput.decoders import decode_message
-from pypgoutput.decoders import ColumnData
-from pypgoutput.decoders import ColumnType
-from pypgoutput.utils import SourceDBHandler, convert_string_to_type, convert_raw_to_change_events
+from pypgoutput.decoders import (
+    PgoutputMessage,
+    Begin,
+    Commit,
+    Origin,
+    Relation,
+    TupleData,
+    Insert,
+    Update,
+    Delete,
+    Truncate,
+    decode_message,
+    ColumnData,
+    ColumnType
+)
+
+from pypgoutput.utils import (
+    SourceDBHandler,
+    convert_string_to_type
+)
+
+from pypgoutput.reader import (
+  LogicalReplicationReader
+)
